# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CATAMS API
  description: |
    Casual Academic Time Allocation Management System (CATAMS) API specification.
    
    This API provides endpoints for managing academic staff time allocation, including:
    - User authentication and management
    - Timesheet creation and management
    - Approval workflow management
    
    The system supports three main roles:
    - **LECTURER**: Can create and approve timesheets for tutors
    - **TUTOR**: Can view their timesheets and request modifications
    - **ADMIN**: Has full system access including HR functions
  version: 1.0.0
  contact:
    name: CATAMS Development Team
    email: dev@catams.edu.au
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.catams.edu.au
    description: Production server
...
paths:
...
  /timesheets:
    get:
      tags:
        - Timesheets
      summary: Get timesheets
      description: |
        Retrieve timesheets based on filters and user permissions.
        
        **Access rules**:
        - **LECTURER**: Can view all timesheets for their courses
        - **TUTOR**: Can view only their own timesheets
        - **ADMIN**: Can view all timesheets in the system
        
        **Filtering**: Results can be filtered by date range, course, tutor, and approval status.
      operationId: getTimesheets
      security:
        - bearerAuth: []
      parameters:
        - name: tutorId
          in: query
          description: Filter by tutor ID (LECTURER and ADMIN only)
          required: false
          schema:
            type: integer
            format: int64
        - name: courseId
          in: query
          description: Filter by course ID
          required: false
          schema:
            type: integer
            format: int64
        - name: startDate
          in: query
          description: Filter by start date (YYYY-MM-DD)
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Filter by end date (YYYY-MM-DD)
          required: false
          schema:
            type: string
            format: date
        - name: status
          in: query
          description: Filter by approval status
          required: false
          schema:
            $ref: '#/components/schemas/ApprovalStatus'
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Page size
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Timesheets retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetPageResponse'
              examples:
                timesheet_list:
                  summary: List of timesheets
                  value:
                    content:
                      - id: 1
                        tutorId: 123
                        tutorName: "Alice Johnson"
                        courseId: 456
                        courseName: "COMP1001 - Introduction to Programming"
                        weekStartDate: "2025-07-28"
                        hours: 10.5
                        hourlyRate: 45.00
                        totalPay: 472.50
                        description: "Tutorial sessions and assignment marking"
                        status: "PENDING_LECTURER_APPROVAL"
                        createdAt: "2025-07-28T09:00:00Z"
                        updatedAt: "2025-07-28T09:00:00Z"
                    pageable:
                      pageNumber: 0
                      pageSize: 20
                      sort:
                        sorted: true
                        orderBy: "weekStartDate"
                        direction: "DESC"
                    totalElements: 45
                    totalPages: 3
                    first: true
                    last: false
                    numberOfElements: 20
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Timesheets
      summary: Create new timesheet
      description: |
        Create a new timesheet entry.
        
        **Required role**: LECTURER, ADMIN
        
        **Business rules**:
        - LECTURER can only create timesheets for tutors in their courses
        - ADMIN can create timesheets for any tutor
        - Week start date must be a Monday
        - Hours must be positive and reasonable (≤ 60 hours per week)
        - Hourly rate must be within configured limits
        - Duplicate timesheets for same tutor/course/week are not allowed
      operationId: createTimesheet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimesheetCreateRequest'
            examples:
              create_timesheet:
                summary: Create timesheet example
                value:
                  tutorId: 123
                  courseId: 456
                  weekStartDate: "2025-07-28"
                  hours: 10.5
                  hourlyRate: 45.00
                  description: "Tutorial sessions and assignment marking"
      responses:
        '201':
          description: Timesheet created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetResponse'
              examples:
                created_timesheet:
                  summary: Successfully created timesheet
                  value:
                    id: 789
                    tutorId: 123
                    tutorName: "Alice Johnson"
                    courseId: 456
                    courseName: "COMP1001 - Introduction to Programming"
                    weekStartDate: "2025-07-28"
                    hours: 10.5
                    hourlyRate: 45.00
                    totalPay: 472.50
                    description: "Tutorial sessions and assignment marking"
                    status: "DRAFT"
                    createdAt: "2025-08-01T14:30:00Z"
                    updatedAt: "2025-08-01T14:30:00Z"
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicate_timesheet:
                  summary: Duplicate timesheet error
                  value:
                    timestamp: "2025-08-01T10:15:30.123Z"
                    status: 400
                    error: "DUPLICATE_TIMESHEET"
                    message: "Timesheet already exists for this tutor/course/week"
                    path: "/api/timesheets"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tutor or course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /timesheets/{id}:
    get:
      tags:
        - Timesheets
      summary: Get specific timesheet
      description: |
        Retrieve a specific timesheet by ID.
        
        **Access rules**:
        - **LECTURER**: Can view timesheets for their courses
        - **TUTOR**: Can view only their own timesheets
        - **ADMIN**: Can view any timesheet
      operationId: getTimesheetById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Timesheet ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Timesheet retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetDetailResponse'
              examples:
                timesheet_detail:
                  summary: Detailed timesheet information
                  value:
                    id: 789
                    tutorId: 123
                    tutorName: "Alice Johnson"
                    tutorEmail: "alice.johnson@university.edu.au"
                    courseId: 456
                    courseName: "COMP1001 - Introduction to Programming"
                    courseCode: "COMP1001"
                    lecturerId: 999
                    lecturerName: "Dr. John Smith"
                    weekStartDate: "2025-07-28"
                    hours: 10.5
                    hourlyRate: 45.00
                    totalPay: 472.50
                    description: "Tutorial sessions and assignment marking"
                    status: "PENDING_LECTURER_APPROVAL"
                    createdAt: "2025-07-28T09:00:00Z"
                    updatedAt: "2025-07-28T09:00:00Z"
                    approvalHistory:
                      - id: 1
                        approverId: 999
                        approverName: "Dr. John Smith"
                        action: "SUBMIT_FOR_APPROVAL"
                        comment: "Initial submission"
                        timestamp: "2025-07-28T09:00:00Z"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Timesheet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Timesheets
      summary: Update timesheet
      description: |
        Update an existing timesheet.
        
        **Access rules**:
        - **LECTURER**: Can update timesheets for their courses (if not yet approved)
        - **ADMIN**: Can update any timesheet (with appropriate workflow constraints)
        - **TUTOR**: Cannot directly update timesheets (must request modifications)
        
        **Business rules**:
        - Cannot update timesheets that are already approved
        - Updates reset approval status to DRAFT
        - All validation rules from creation apply
      operationId: updateTimesheet
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Timesheet ID
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimesheetUpdateRequest'
            examples:
              update_timesheet:
                summary: Update timesheet example
                value:
                  hours: 12.0
                  hourlyRate: 45.00
                  description: "Tutorial sessions, assignment marking, and consultation hours"
      responses:
        '200':
          description: Timesheet updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetResponse'
              examples:
                updated_timesheet:
                  summary: Successfully updated timesheet
                  value:
                    id: 789
                    tutorId: 123
                    tutorName: "Alice Johnson"
                    courseId: 456
                    courseName: "COMP1001 - Introduction to Programming"
                    weekStartDate: "2025-07-28"
                    hours: 12.0
                    hourlyRate: 45.00
                    totalPay: 540.00
                    description: "Tutorial sessions, assignment marking, and consultation hours"
                    status: "DRAFT"
                    createdAt: "2025-07-28T09:00:00Z"
                    updatedAt: "2025-08-01T15:45:00Z"
        '400':
          description: Invalid request data or business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                already_approved:
                  summary: Cannot update approved timesheet
                  value:
                    timestamp: "2025-08-01T10:15:30.123Z"
                    status: 400
                    error: "INVALID_STATE_TRANSITION"
                    message: "Cannot update timesheet that has been approved"
                    path: "/api/timesheets/789"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Timesheet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Timesheets
      summary: Delete timesheet
      description: |
        Delete an existing timesheet.
        
        **Access rules**:
        - **LECTURER**: Can delete timesheets for their courses (if not yet approved)
        - **ADMIN**: Can delete any timesheet (with appropriate workflow constraints)
        - **TUTOR**: Cannot directly delete timesheets
        
        **Business rules**:
        - Cannot delete timesheets that are already approved
        - Cannot delete timesheets that have been submitted to HR
        - Deletion creates audit log entry
        - Related approval history is preserved for audit purposes
      operationId: deleteTimesheet
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Timesheet ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Timesheet deleted successfully
        '400':
          description: Cannot delete timesheet due to business rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                cannot_delete_approved:
                  summary: Cannot delete approved timesheet
                  value:
                    timestamp: "2025-08-01T10:15:30.123Z"
                    status: 400
                    error: "INVALID_STATE_TRANSITION"
                    message: "Cannot delete timesheet that has been approved"
                    path: "/api/timesheets/789"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Timesheet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    TimesheetCreateRequest:
      type: object
      required:
        - tutorId
        - courseId
        - weekStartDate
        - hours
        - hourlyRate
        - description
      properties:
        tutorId:
          type: integer
          format: int64
          description: ID of the tutor this timesheet is for
          example: 123
        courseId:
          type: integer
          format: int64
          description: ID of the course this work relates to
          example: 456
        weekStartDate:
          type: string
          format: date
          description: Start date of the work week (must be a Monday)
          example: "2025-07-28"
        hours:
          type: number
          format: decimal
          minimum: 0.1
          maximum: 60.0
          description: Number of hours worked
          example: 10.5
        hourlyRate:
          type: number
          format: decimal
          minimum: 0.01
          maximum: 200.00
          description: Hourly rate in AUD
          example: 45.00
        description:
          type: string
          description: Description of work performed
          example: "Tutorial sessions and assignment marking"
          maxLength: 1000
...
    TimesheetPageResponse:
      type: object
      required:
        - content
        - pageable
        - totalElements
        - totalPages
        - first
        - last
        - numberOfElements
      properties:
        content:
          type: array
          description: List of timesheets on this page
          items:
            $ref: '#/components/schemas/TimesheetResponse'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
          format: int64
          description: Total number of timesheets matching the query
          example: 45
        totalPages:
          type: integer
          description: Total number of pages
          example: 3
        first:
          type: boolean
          description: Whether this is the first page
          example: true
        last:
          type: boolean
          description: Whether this is the last page
          example: false
        numberOfElements:
          type: integer
          description: Number of elements on this page
          example: 20
    Pageable:
      type: object
      required:
        - pageNumber
        - pageSize
      properties:
        pageNumber:
          type: integer
          description: Current page number (0-based)
          example: 0
        pageSize:
          type: integer
          description: Page size
          example: 20
        sort:
          $ref: '#/components/schemas/Sort'
...
    ApprovalStatus:
      type: string
      enum:
        - DRAFT
        - PENDING_LECTURER_APPROVAL
        - APPROVED
        - REJECTED
        - MODIFICATION_REQUESTED
        - PENDING_HR_REVIEW
        - FINAL_APPROVED
      description: |
        Approval status of a timesheet:
        - **DRAFT**: Initial state, being prepared
        - **PENDING_LECTURER_APPROVAL**: Submitted for lecturer review
        - **APPROVED**: Approved by lecturer, ready for HR
        - **REJECTED**: Rejected by lecturer or HR
        - **MODIFICATION_REQUESTED**: Lecturer requested changes
        - **PENDING_HR_REVIEW**: Awaiting final HR approval
        - **FINAL_APPROVED**: Final approval by HR, ready for payment
      example: "PENDING_LECTURER_APPROVAL"
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
          example: "2025-08-01T10:15:30.123Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error code for programmatic handling
          example: "VALIDATION_FAILED"
        message:
          type: string
          description: Human-readable error message
          example: "Validation failed: email: Email format is invalid"
        path:
          type: string
          description: Request path that caused the error
          example: "/api/auth/login"

# Security configuration
security:
  - bearerAuth: []